requirements.txt
plotly
Pillow
sqlalchemy
flask
requests
Jinja2
pytest
black
flask-cors
streamlit
pandas
openai
python-dotenv
reportlab
matplotlib
fastapi
uvicorn
python-multipart
openpyxl
langchain
fpdf
langchain_community